# Log::Metadata optimizations