--- a/spec/std/log/metadata_spec.cr
+++ b/spec/std/log/metadata_spec.cr
@@ -74,6 +74,22 @@ describe Log::Metadata do
     md.@overriden_size.should eq(1)
     md.@parent.should be_nil
   end
+
+  it "[]" do
+    md = m({a: 1, b: 2}).extend({a: 3})
+
+    md[:a].should eq(3)
+    md[:b].should eq(2)
+    expect_raises(KeyError) { md[:c] }
+  end
+
+  it "[]?" do
+    md = m({a: 1, b: 2}).extend({a: 3})
+
+    md[:a]?.should eq(3)
+    md[:b]?.should eq(2)
+    md[:c]?.should be_nil
+  end
 end
 
 describe Log::Metadata::Value do