--- a/src/log/entry.cr
+++ b/src/log/entry.cr
@@ -43,6 +43,5 @@ struct Log::Entry
   getter exception : Exception?
 
   def initialize(@source : String, @severity : Severity, @message : String, @data : Log::Metadata, @exception : Exception?)
-    raise ArgumentError.new "Expected hash data, not #{data.raw.class}" unless data.raw.as?(Hash)
   end
 end