--- a/src/log/json.cr
+++ b/src/log/json.cr
@@ -9,6 +9,21 @@ class Log::Metadata
   # log_entry.context.to_json # => "{\"user_id\":1}"
   # ```
   def to_json(builder : JSON::Builder) : Nil
-    @raw.to_json builder
+    builder.object do
+      each do |(key, value)|
+        builder.field key.to_json_object_key do
+          value.to_json(builder)
+        end
+      end
+    end
+  end
+
+  struct Value
+    # Returns `Log::Metadata::Value` as JSON value.
+    #
+    # NOTE: `require "log/json"` is required to opt-in to this feature.
+    def to_json(builder : JSON::Builder) : Nil
+      @raw.to_json builder
+    end
   end
 end