--- a/src/log/metadata.cr
+++ b/src/log/metadata.cr
@@ -20,8 +20,10 @@ class Log::Metadata
   # When the metadata is defragmented max_total_size will be updated with size
   protected getter max_total_size : Int32
   @max_total_size = uninitialized Int32
+  # How many entries are potentially overriden from parent (ie: initial entries.size)
+  @overriden_size = uninitialized Int32
   # How many entries are stored from @first.
-  # Initially are the one explictly overriden in entries argument.
+  # Initially are @overriden_size, the one explictly overriden in entries argument.
   # When the metadata is defragmented @size will be increased up to
   # the actual number of entries resulting from merging the parent
   @size = uninitialized Int32
@@ -36,7 +38,7 @@ class Log::Metadata
   end
 
   protected def setup(@parent : Metadata?, entries : NamedTuple | Hash)
-    @size = entries.size
+    @size = @overriden_size = entries.size
     @max_total_size = @size + (@parent.try(&.max_total_size) || 0)
     ptr_entries = pointerof(@first)
 
@@ -80,18 +82,24 @@ class Log::Metadata
 
   # Removes the reference to *parent*. Flattening the entries from it into *self*.
   # *self* was originally allocated with enough entries to perform this action.
+  #
+  # If multiple threads execute defrag concurrently, the entries
+  # will be recomputed, but the result should be the same.
+  #
+  # * @parent.nil? signals if the defrag is needed/done
+  # * The values of @overriden_size, pointerof(@first) are never changed
+  # * @parent is set at the very end of the method
   protected def defrag
     parent = @parent
     return if parent.nil?
 
-    total_size = @size
-    overriden_entries_size = @size
+    total_size = @overriden_size
     ptr_entries = pointerof(@first)
-    next_free_entry = ptr_entries + overriden_entries_size
+    next_free_entry = ptr_entries + @overriden_size
 
     parent.each do |(key, value)|
       overriden = false
-      overriden_entries_size.times do |i|
+      @overriden_size.times do |i|
         if ptr_entries[i][:key] == key
           overriden = true
           break