Related to https://github.com/crystal-lang/crystal/pull/9227#issuecomment-626840159, this PR introduces some modifications proposed in #9188 and some optimizations needed to avoid performance penalties in the presence of implicit contexts mainly.

This PR effectively 
* changes the representation of Metadata to be a structure for the top-level (`Log::Metadata`) and the datum (`Log::Metadata::Value`) for the rest of it
* `Log::Metadata` keys are symbols, but `Log::Metadata::Value` keys are still String. (main api `Log.context.set v: expr` is not changed)
* Since Metadata is always a hash-like structure, there is no need for runtime validation 🚀  
* Metadata are no longer _mergeable_ but they are _extendable_ which was the original intention.
  * This API allow a better representation that delays the merging when it's needed 🚀  
* The output of the metadata (context and local data) is tweaked a bit to be less noisy
* Finally the `Log::Metadata::Value` is no longer immutable. If we would have a way to freeze objects it would be better story. But cloning, always, is way too expensive.

cc: @paulcsmith @carlhoerberg