@wontruefree 2020-05-14
I'm curious did you do any benchmarking with this?