@waj 2020-05-14
> This is some whacky performance optimization for this early in the game

It's not. There are users already using Crystal in production with high demanding apps. Adding logging shouldn't be a bottleneck for them. And we're really interested they try these new apis.