@asterite 2020-05-14
I also remember kemal benchmarks changed a lot depending on whether logging was turned on or off. So optimizing logging as much as possible is great!