@RX14 2020-05-14
"Whacky" wasn't meant as an insult...