@wontruefree 2020-05-14
I am not saying the optimizations are wacky.  It is just when something is done for performance sometime things can be surprising and it is nice to measure before and after.  It is kind of like TDD for optimizations.