@bcardiff 2020-05-14
We've made different benchmarks, to meassure the cost of the Log::Metadata by (by using a NullBackend), and to measure the output (using IOBackend with a NullIO). Or how does it impact in the performance on an HTTP server.

There are many things that can be compared. You can view some of those intermediate benchmarks at https://gist.github.com/bcardiff/c4d1d3fcd92e0b804d518af8dbf314fa

The basis is that, by not paying the cost of the merge upon creation, you can create and discard better contexts that are not used.

```cr
require "log"
require "benchmark"

Benchmark.ips do |x|
  x.report("log-fiber-context") { Log.with_context { Log.context.set(a: 1) } }
end
```

|   |  result  |
|--|--|
| Without PR | log-fiber-context   3.69M (271.08ns) (± 5.41%)  561B/op |
| With PR | log-fiber-context  17.14M ( 58.34ns) (± 4.30%)  80.0B/op |

Of course, there are some pathological cases. You can chain lots and lots of context and never release them. Those will play worse in this PR as is, but they are pathological cases. Yet the changes in the API allows further improvements without breaking changes: like allocating the struct differently, or forcing defrag based con some conditions to limit the metadata chain size.