@wontruefree 2020-05-14
Thanks for the update on the benchmarks.

This is super interesting to look into it looks like there are gains across the board.  Some are < 10% increases and some are really large.

I think it I can more understand the performance impacts now.