@straight-shoota 2020-05-15
> I also remember kemal benchmarks changed a lot depending on whether logging was turned on or off. So optimizing logging as much as possible is great!

AFAIK that performance impact is caused by having logging vs. no logging at all and is probably mostly limited by the backend (typically file IO). The implementation of the logging framework is unlikely to have such a huge impact. I mean `no logging vs. logging` is a much bigger difference than `non-optimized vs. optimized logging implementation`.
It's still worth optimizing the implementation, but I doubt it makes kemal benchmarks with logging significantly faster.