@didactic-drunk 2020-05-15
Results from a WIP program with lots of structured log statements.

Before:
```
LOG_LEVEL=error  0.02s # No logging
LOG_LEVEL=notice 0.10s
LOG_LEVEL=info   0.99s
LOG-LEVEL=debug  2.16s
```

After:
```
LOG_LEVEL=error   0.02s # No logging
LOG_LEVEL=notice  0.09s
LOG_LEVEL=info    0.78s
LOG-LEVEL=debug   1.31s
```

Runtimes are ~+|-5% probably because the system isn't idle.

@straight-shoota is correct with the impact of no logging vs logging having a greater impact.  

@bcardiff Faster structured logging is still necessary and I hope the optimizations go further.